pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "aip-tumor"
version = "0.1.0"
description = "Brain tumor classification + segmentation inference (FastAPI + optional CLI)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "numpy>=1.26.0",
    "pillow>=10.0.0",
    "python-multipart>=0.0.9",
    "torch>=2.2.0",
    "torchvision>=0.17.0",
    "typer>=0.12.0",
    "uvicorn[standard]>=0.30.0",
]

[project.optional-dependencies]
train = [
    "matplotlib>=3.8.0",
    "pandas>=2.0.0",
    "scikit-learn>=1.3.0",
    "kagglehub>=0.2.0",
]

[project.scripts]
tumor-cli = "src.tumor_app.cli:app"
aip-train = "src.training.train_models:main"

[tool.setuptools]
packages = [
    "src",
    "src.models",
    "src.training",
    "src.data",
    "src.utils",
    "src.tumor_app",
]

[tool.setuptools.package-data]
"src.tumor_app" = ["static/*.html"]
